Moved worker for primary color to a new customHook in a new file
