+++ src/components/RenderPrimaryColors.jsx
@@ -1,36 +1,13 @@
-import { useEffect, useState, useRef } from 'react';
 import IndividualPrimaryColor from './IndividualPrimaryColor';
-
-function useWorkerPrimaryColors() {}
+import useWorkerPrimaryColors from '../hooks/useWorkerPrimaryColors';
 
 export default function RenderPrimaryColors(props) {
 	const { avgColors, colorMode, colorTolerance, divsQty } = props;
-	const [colorList, setColorList] = useState([]);
-	const [isLoading, setIsLoading] = useState(false);
-
-	const workerRef = useRef(
-		new Worker('../src/workers/workerPrimaryColors.js', { type: 'module' })
+	const { colorList, isLoading } = useWorkerPrimaryColors(
+		avgColors,
+		colorMode,
+		colorTolerance
 	);
-
-	// Get the current state of worker, and save in a variable
-	const worker = workerRef.current;
-
-	useEffect(() => {
-		setIsLoading(true);
-		worker.postMessage({
-			msg: 'get-primary-colors',
-			avgColors,
-			colorMode,
-			colorTolerance,
-		});
-	}, [avgColors, colorMode, colorTolerance]);
-
-	worker.onmessage = (e) => {
-		if (e.data.primaryColor) {
-			setColorList(e.data.primaryColor);
-			setIsLoading(false);
-		}
-	};
 
 	return (
 		<div className='my-10'>
